修复流未关闭问题
