--- src/main/java/cc/mrbird/common/util/AddressUtils.java
@@ -1,61 +1,0 @@
-package cc.mrbird.common.util;
-
-import org.apache.commons.io.FileUtils;
-import org.lionsoul.ip2region.DataBlock;
-import org.lionsoul.ip2region.DbConfig;
-import org.lionsoul.ip2region.DbSearcher;
-import org.lionsoul.ip2region.Util;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.File;
-import java.lang.reflect.Method;
-
-public class AddressUtils {
-
-    private AddressUtils() {
-    }
-
-    private static Logger log = LoggerFactory.getLogger(AddressUtils.class);
-
-    public static String getCityInfo(String ip) {
-        try {
-            String dbPath = AddressUtils.class.getResource("/ip2region/ip2region.db").getPath();
-            File file = new File(dbPath);
-            if (!file.exists()) {
-                String tmpDir = System.getProperties().getProperty("java.io.tmpdir");
-                dbPath = tmpDir + "ip.db";
-                file = new File(dbPath);
-                FileUtils.copyInputStreamToFile(AddressUtils.class.getClassLoader().getResourceAsStream("classpath:ip2region/ip2region.db"), file);
-            }
-            int algorithm = DbSearcher.BTREE_ALGORITHM;
-            DbConfig config = new DbConfig();
-            DbSearcher searcher = new DbSearcher(config, file.getPath());
-            Method method = null;
-            switch (algorithm) {
-                case DbSearcher.BTREE_ALGORITHM:
-                    method = searcher.getClass().getMethod("btreeSearch", String.class);
-                    break;
-                case DbSearcher.BINARY_ALGORITHM:
-                    method = searcher.getClass().getMethod("binarySearch", String.class);
-                    break;
-                case DbSearcher.MEMORY_ALGORITYM:
-                    method = searcher.getClass().getMethod("memorySearch", String.class);
-                    break;
-                default:
-                    method = searcher.getClass().getMethod("memorySearch", String.class);
-                    break;
-            }
-            DataBlock dataBlock = null;
-            if (!Util.isIpAddress(ip)) {
-                log.error("Error: Invalid ip address");
-            }
-            dataBlock = (DataBlock) method.invoke(searcher, ip);
-            return dataBlock.getRegion();
-        } catch (Exception e) {
-            log.error("获取地址信息异常：{}", e);
-        }
-        return "";
-    }
-
-}
